Removed profiles and split users into two models.
The logic behind this may seem anti-DRY, but I'm just trying to make things a bit easier. Originally I was going to use a polymorphic setup, however for simplicity I am keeping the User and Patient models/tables separate. In the past, I've seen multiple other databases that use this pattern, and the separation of concerns makes observing the data that much better.